scraper_service/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
pandas
python-dotenv
requests
beautifulsoup4
tls-client
markdownify
regex
numpy
